start of cost calc

+++ src/tui.rs
@@ -415,16 +415,18 @@ where
                         },
                         KeyCode::Up => session.scroll_state.scroll_up(),
                         KeyCode::Down => session.scroll_state.scroll_down(),
-                        KeyCode::Backspace => if let PopupState::TodoList(list_state) = &session.popup_state {
-                            let index = list_state.selected().unwrap_or_default();
-                            let mut todo = session.runner.context.todo_list.lock().await;
+                        KeyCode::Backspace => {
+                            if let PopupState::TodoList(list_state) = &session.popup_state {
+                                let index = list_state.selected().unwrap_or_default();
+                                let mut todo = session.runner.context.todo_list.lock().await;
 
-                            if let Some(key) =
-                                todo.iter().nth(index).map(|(key, _)| key.clone())
-                            {
-                                todo.remove_entry(&key);
+                                if let Some(key) =
+                                    todo.iter().nth(index).map(|(key, _)| key.clone())
+                                {
+                                    todo.remove_entry(&key);
+                                }
                             }
-                        },
+                        }
                         KeyCode::Enter => match &session.popup_state {
                             PopupState::ModelSelect(list_state) => {
                                 let index = list_state.selected().unwrap_or_default();
@@ -537,7 +539,7 @@ where
                     };
                 }
                 TuiEvent::Exit => {
-                    session.runner.shutdown();
+                    session.runner.cancel();
                     session.save().await.unwrap();
                     break;
                 }
